[REFACTOR] 카테고리 기능
불필요한 주석 제거, 패키지 구조 변경
