+++ honeybread-mgmtadmin/src/main/java/com/whatsub/honeybread/mgmtadmin/domain/category/CategoryQueryService.java
@@ -3,6 +3,7 @@ package com.whatsub.honeybread.mgmtadmin.domain.category;
 import com.whatsub.honeybread.core.domain.category.CategoryRepository;
 import com.whatsub.honeybread.core.infra.errors.ErrorCode;
 import com.whatsub.honeybread.core.infra.exception.HoneyBreadException;
+import com.whatsub.honeybread.mgmtadmin.domain.category.dto.CategoryResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
